Fixed documentation and added a script to validate config files

+++ project/scripts/reset-prod.sh
@@ -2,6 +2,12 @@
 
 ROOT_DIR=$(git rev-parse --show-toplevel)
 
+./project/scripts/validate-prod-config.sh
+if test ! "$?" -eq "0"; then
+    echo "Check the documentation at project/README.md to define configurations for all containers"
+    exit "$?"
+fi
+
 cd "$ROOT_DIR"/project || exit
 
 docker compose down
